Move user login logic from controller to service
