+++ controllers/user.controller.js
@@ -1,8 +1,4 @@
 /** @typedef {import("../helpers/types").RequestHandler} RequestHandler */
-const jwt = require("jsonwebtoken");
-const settings = require("../helpers/settings");
-const messages = require("../helpers/messages");
-const { UnauthorizedError } = require("../helpers/errors");
 const { filterObj } = require("../helpers/tools");
 
 /** @type {RequestHandler} */
@@ -13,25 +9,10 @@ async function signup(req, res) {
 
 /** @type {RequestHandler} */
 async function signin(req, res) {
-  const credentials = req.body;
-  const user = await req.services.user.getByEmail(credentials.email);
-  if (!(await user.comparePassword(credentials.password))) {
-    throw new UnauthorizedError(messages.users.loginError);
-  }
-
-  const payload = filterObj(user, [
-    ["_id", "id"],
-    "email",
-    "subscription",
-    "avatarURL",
-  ]);
-
-  const token = jwt.sign(payload, settings.authentication.jwtSecret, {
-    expiresIn: settings.authentication.jwtLifetime,
-  });
-  await req.services.user.updateById(user.id, { token });
+  const { email, password } = req.body;
+  const user = await req.services.user.login(email, password);
   res.json({
-    token,
+    token: user.token,
     user: filterObj(user, ["email", "subscription", "avatarURL"]),
   });
 }
@@ -39,10 +20,7 @@ async function signin(req, res) {
 /** @type {RequestHandler} */
 async function signout(req, res) {
   await req.services.user.updateById(req.user.id, { token: null });
-
-  // Clearing the req.user just to be on the safer side
   delete req.user;
-
   res.status(204).end();
 }
